chore(navigation): remove legacy service (#1051)
Removing the legacy `services/navigation` to simplify navigation
implementation.

See commits for sequence.

[Asana](https://app.asana.com/0/1204008699308084/1209189437518189/f)
